Separate on-chain / off-chain code in the ContractSM module. Moved off-chain into Contract/PAB
